// frontend/src/app/models/course-management/coursepublicMainData.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { FinanceImportLogData } from './../financial-management';

export interface CoursepublicMainData {
    /** สถานะการใช้งาน */
    activeFlag?: boolean;
    /** ให้บริการบัวศรีไอดี */
    buasriStatus?: boolean;
    /** รูปแบบการให้ประกาศนียบัตร 30006000 */
    certificateFormat?: number[];
    /** ส่วนแบ่งศูนย์บริการวิชาการ */
    costShareCenterPercent?: number;
    /** ส่วนแบ่งคณะ */
    costShareDepPercent?: number;
    /** ส่วนแบ่งมหาวิทยาลัย */
    costShareGlobalPercent?: number;
    /** Count Actived */
    countActived?: number;
    /** วันที่เรียนสิ้นสุด */
    courseClassEnd?: Date;
    /** วันที่เรียนเริ่มต้น */
    courseClassStart?: Date;
    /** custom รหัสหลักสูตร */
    courseCode?: string;
    /** รูปแบบการเรียนการสอน */
    courseFormat?: number;
    /** รายละเอียดรูปแบบการสอนเพิ่มเติม (ภาษาอังกฤษ) */
    courseFormatDescEn?: string;
    /** รายละเอียดรูปแบบการสอนเพิ่มเติม (ภาษาไทย) */
    courseFormatDescTh?: string;
    /** รุ่นที่ */
    courseGeneration?: number;
    /** FK (course_main) */
    courseId?: number;
    /** วันที่โปรโมทสิ้นสุด */
    coursePublicEnd?: Date;
    /** PK */
    coursepublicId?: number;
    /** วันที่โปรโมทเริ่มต้น */
    coursePublicStart?: Date;
    /** สถานะเปิดรอบหลักสูตร 30014000 */
    coursepublicStatus?: number;
    /** custom สถานะเปิดรอบหลักสูตร (ภาษาอังกฤษ) */
    coursepublicStatusEn?: string;
    /** custom สถานะเปิดรอบหลักสูตร (ภาษาไทย) */
    coursepublicStatusTh?: string;
    /** วันที่ปิดรับสมัคร */
    courseRegisEnd?: Date;
    /** วันที่เปิดรับสมัคร */
    courseRegisStart?: Date;
    /** เวลาเรียน (ภาษาอังกฤษ) */
    courseTimeEn?: string;
    /** เวลาเรียน (ภาษาไทย) */
    courseTimeTh?: string;
    createById?: number;
    createDate?: string;
    /** คณะ/หน่วยงาน */
    depIdLevel1?: number;
    /** ภาควิชา */
    depIdLevel2?: number;
    /** ค่าลงทะเบียน / ราคา */
    feeAmount?: number;
    /** การคิดค่าลงทะเบียน 30008000 */
    feeStatus?: number;
    /** custom (coursepublic_media) filename */
    filename?: string;
    financeImportLog?: FinanceImportLogData;
    first?: number;
    /** grade format */
    gradeFormat?: number;
    /** จำนวนที่เปิดรับ */
    maxEnroll?: number;
    /** custom ชื่อ Thumbnail (ภาษาอังกฤษ) (coursepublic_media) */
    mediaNameEn?: string;
    /** custom ชื่อ Thumbnail (ภาษาไทย) (coursepublic_media) */
    mediaNameTh?: string;
    mode?: string;
    moodleCourseId?:number;
    /** custom (coursepublic_media) Module */
    module?: number;
    /** course type name en */
    nameEn?: string;
    /** course type name local */
    nameLo?: string;
    /** Popular */
    popularStatus?: boolean;
    /** custom (coursepublic_media) Prefix */
    prefix?: string;
    /** ราคาโปรโมชั่น */
    promotionAmount?: number;
    /** วันสุดท้ายที่จัดโปรโมชั่นส่วนลด */
    promotionEnd?: string;
    /**  วันที่เริ่มจัดโปรโมชั่นส่วนลด */
    promotionStart?: string;
    /** รูปแบบการเปิดหลักสูตร 30007000 */
    publicFormat?: number;
    /** ชื่อรอบหลักสูตร (ภาษาอังกฤษ) */
    publicNameEn?: string;
    /** ชื่อรอบหลักสูตร (ภาษาไทย) */
    publicNameTh?: string;
    /** เผยแพร่บนเว็บไซต์ */
    publicStatus?: boolean;
    rowNum?: number;
    size?: number;
    /** custom thumbnail */
    thumbnail?: string;
    imgSrc?: string;
    memberCount?:number;
}